requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
fastapi==0.95.2
frozenlist==1.3.3
h11==0.14.0
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
pydantic==1.10.8
pymongo==4.3.3
python-dotenv==0.21.0
pytz==2023.3
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.1
urllib3==2.0.2
uvicorn==0.22.0
yarl==1.9.2
